+++ packages/shared/bundler-shared-compat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pie-element/bundler-shared
 
+## 0.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [8d69fb5]
+  - @pie-element/element-bundler@0.1.2-next.0
+
 ## 0.1.1
 
 ### Patch Changes